requirements.txt
Flask==2.3.3
requests==2.31.0
psycopg2==2.9.7
schedule==1.2.0
pytz==2023.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0